chore: move PoolKey to utils/pool_key so the async provider can share it
